code file refactoring, make classes as independent files
