+++ ReoGrid/Views/Header/SpaceView.cs
@@ -1,0 +1,46 @@
+﻿/*****************************************************************************
+ * 
+ * ReoGrid - .NET Spreadsheet Control
+ * 
+ * https://reogrid.net/
+ *
+ * THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY
+ * KIND, EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR
+ * PURPOSE.
+ *
+ * Author: Jingwood <jingwood at unvell.com>
+ *
+ * Copyright (c) 2012-2021 Jingwood, unvell.com, all rights reserved.
+ * 
+ ****************************************************************************/
+
+using System;
+
+#if WINFORM || ANDROID
+using RGFloat = System.Single;
+#elif WPF || iOS
+using RGFloat = System.Double;
+#endif
+
+using unvell.Common;
+
+using unvell.ReoGrid.Actions;
+using unvell.ReoGrid.Events;
+using unvell.ReoGrid.Graphics;
+using unvell.ReoGrid.Rendering;
+using unvell.ReoGrid.Interaction;
+using unvell.ReoGrid.Main;
+
+namespace unvell.ReoGrid.Views
+{
+	class SpaceView : View
+	{
+		public SpaceView() : base() { }
+
+		public override void Draw(CellDrawingContext dc)
+		{
+			dc.Graphics.FillRectangle(bounds, Rendering.StaticResources.SystemColor_Control);
+		}
+	}
+}
